rename `trait Shard` -> `trait CongestionStrategy`
as pointed out in the review, the name `Shard` made very little sense
