+++ tools/congestion-model/src/evaluation/summary_table.rs
@@ -3,11 +3,11 @@ use crate::{Model, PGAS};
 pub fn print_summary_header() {
     println!(
         "{:<25}{:<25}{:>25}{:>25}{:>25}",
-        "WORKLOAD", "DESIGN", "BURNT GAS", "TRANSACTIONS FINISHED", "MAX QUEUE LEN",
+        "WORKLOAD", "STRATEGY", "BURNT GAS", "TRANSACTIONS FINISHED", "MAX QUEUE LEN",
     );
 }
 
-pub fn print_summary_row(model: &Model, workload: &str, design: &str) {
+pub fn print_summary_row(model: &Model, workload: &str, strategy: &str) {
     let queues = model.queue_lengths();
     let throughput = model.gas_throughput();
     let progress = model.progress();
@@ -19,7 +19,7 @@ pub fn print_summary_row(model: &Model, workload: &str, design: &str) {
     }
 
     println!(
-        "{workload:<25}{design:<25}{:>20} PGas{:>25}{:>25}",
+        "{workload:<25}{strategy:<25}{:>20} PGas{:>25}{:>25}",
         throughput.total / PGAS,
         progress.finished_transactions,
         max_queue_len
