+++ tools/congestion-model/src/strategy/mod.rs
@@ -9,9 +9,13 @@ mod global_tx_stop;
 mod no_queues;
 mod simple_backpressure;
 
-/// Implement the shard behavior to define a new congestion control design.
-pub trait Shard {
-    /// Initial state and register all necessary queues.
+/// Implement the shard behavior to define a new congestion control strategy.
+///
+/// The model execution will take one `CongestionStrategy` trait object per
+/// shard. They normally are all of the same concrete type and hence execute
+/// the same code on each shard.
+pub trait CongestionStrategy {
+    /// Initial state and register all necessary queues for one shard.
     fn init(&mut self, id: ShardId, other_shards: &[ShardId], queue_factory: &mut dyn QueueFactory);
 
     /// Decide which receipts to execute, which to delay, and which to forward.
